+++ src/main/java/noobanidus/mods/lootr/blocks/LootrChestBlock.java
@@ -22,6 +22,7 @@ import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
 import noobanidus.mods.lootr.data.NewChestData;
 import noobanidus.mods.lootr.init.ModTiles;
+import noobanidus.mods.lootr.tiles.ILootTile;
 import noobanidus.mods.lootr.tiles.SpecialLootChestTile;
 import noobanidus.mods.lootr.util.ChestUtil;
 
@@ -41,7 +42,9 @@ public class LootrChestBlock extends ChestBlock {
 
   @Override
   public void onReplaced(BlockState oldState, World world, BlockPos pos, BlockState newState, boolean isMoving) {
-    NewChestData.deleteLootChest(world, pos);
+    if (oldState.getBlock() != newState.getBlock()) {
+      NewChestData.deleteLootChest(world, pos);
+    }
     super.onReplaced(oldState, world, pos, newState, isMoving);
   }
